Broke the search portions of Main into their own folder.
